Move mutation observer out into separate repo. Restructure repo to better reflect code organisation and meaning.
