+++ gulpfile.js
@@ -1,5 +1,3 @@
-'use strict';
-
 var gulpAutoTask = require('gulp-auto-task');
 
 gulpAutoTask('{*,**/*}.js', {
