+++ src/global/document-observer.js
@@ -2,9 +2,9 @@ import apiWatch from '../api/watch';
 import attached from '../lifecycle/attached';
 import created from '../lifecycle/created';
 import detached from '../lifecycle/detached';
-import globals from '../globals';
+import globals from './vars';
 import ignored from '../util/ignored';
-import registry from '../registry';
+import registry from './registry';
 import walkTree from '../util/walk-tree';
 
 var DocumentFragment = window.DocumentFragment;
