+++ src/global/vars.js
