+++ src/index.js
@@ -19,10 +19,10 @@ import created from './lifecycle/created';
 import debounce from './util/debounce';
 import defaults from './defaults';
 import detached from './lifecycle/detached';
-import documentObserver from './polyfill/document-observer';
-import elementConstructor from './polyfill/element-constructor';
-import polyfillBinding from './polyfill/binding';
-import registry from './registry';
+import documentObserver from './global/document-observer';
+import elementConstructor from './util/element-constructor';
+import polyfillBinding from './type/custom-element-spec';
+import registry from './global/registry';
 import supportsCustomElements from './support/custom-elements';
 import walkTree from './util/walk-tree';
 import validCustomElement from './support/valid-custom-element';
