+++ src/type/custom-element-spec.js
