+++ src/types.js
@@ -1,0 +1,5 @@
+import customElementSpec from './type/custom-element-spec';
+
+export default {
+  'custom-element-spec': customElementSpec
+};
